services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
qdrant-client==1.16.2
redis==5.0.1
langchain==0.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
sentence-transformers==2.2.2